pyproject.toml
[project]
name = "advanced-ai"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "jupyter>=1.1.1",
    "jupyterlab>=4.5.3",
    "kagglehub>=0.4.2",
    "matplotlib>=3.10.8",
    "notebook>=7.5.3",
    "pandas>=3.0.0",
    "polars>=1.38.0",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "tensorflow>=2.20.0",
    "torchvision>=0.25.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "ty>=0.0.21",
]
